create audio uploading

+++ Backend/demo/src/main/java/com/example/demo/repository/userRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.User;
 
 @Repository
-public interface userRepository extends JpaRepository<User, Integer> {
+public interface userRepository extends JpaRepository<User, Long> {
 	boolean existsByEmail(String email);
 	 User findByEmail(String email);
 }
